rendered all topics with rtk query
