+++ client/src/components/navbar/sortquestions.jsx
@@ -1,24 +1,15 @@
 import '../../styles/navbar/navbar.scss';
-import { useSelector, useDispatch } from 'react-redux';
-import { toggleType } from '../../features/questionType';
+import { useDispatch } from 'react-redux';
+import { updateQType } from '../../features/questionTypeSlice';
 
 export const SortQuestions = () => {
   const dispatch = useDispatch();
-  const questionType = useSelector((state) => state.questionType);
 
   return (
     <>
       <ul className="sort-questions">
-        <li
-          onClick={() => dispatch(toggleType())}
-          id={questionType === 'questions' ? 'questions' : null}>
-          Questions
-        </li>
-        <li
-          onClick={() => dispatch(toggleType())}
-          id={questionType === 'syntax' ? 'syntax' : null}>
-          Syntax
-        </li>
+        <li onClick={() => dispatch(updateQType('questions'))}>Questions</li>
+        <li onClick={() => dispatch(updateQType('syntax'))}> Syntax</li>
       </ul>
     </>
   );
