+++ client/src/components/searchquestions.jsx
@@ -1,7 +1,8 @@
 import { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useDebounce } from 'use-debounce';
-import { getQuestions } from '../features/questions';
+//import { questionsApi } from '../services/questionsApi';
+// { useDebounce } from 'use-debounce';
+//import { getQuestions } from '../features/questionsSlice';
 
 // eslint-disable-next-line react/prop-types
 export const SearchQuestions = ({ selectedTopic }) => {
@@ -13,28 +14,29 @@ export const SearchQuestions = ({ selectedTopic }) => {
     console.log(e.target.value);
   };
 
-  const dispatch = useDispatch();
+  //const dispatch = useDispatch();
 
-  const questiontype = useSelector((state) => state.questionType);
+  // const questiontype = useSelector((state) => state.questionType);
+  //const{data, isFetching, error}= questionsApi.endpoints.
 
   const handleSearchSubmit = async (e) => {
-    if (!search) {
-      return dispatch(getQuestions(selectedTopic));
-    }
-    e.preventDefault();
-    try {
-      const data = await fetch(
-        `/${selectedTopic}/search/${questiontype}/${search}`
-      );
-      const response = await data.json();
-      dispatch(getQuestions());
-      //   if (response.length === 0) {
-      //     return setNoData(true);
-      //   }
-      //   setNoData(false);
-    } catch (error) {
-      console.error(error);
-    }
+    //if (!search) {
+    //  return dispatch(getQuestions(selectedTopic));
+    // }
+    // e.preventDefault();
+    // try {
+    //  const data = await fetch(
+    //   `/${selectedTopic}/search/${questiontype}/${search}`
+    // );
+    //const response = await data.json();
+    // dispatch(getQuestions());
+    //   if (response.length === 0) {
+    //     return setNoData(true);
+    //   }
+    //   setNoData(false);
+    // } catch (error) {
+    // console.error(error);
+    //}
   };
   return (
     <form onSubmit={handleSearchSubmit}>
