+++ client/src/components/viewquestions.jsx
@@ -1,16 +1,16 @@
 import { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { getQuestions } from '../features/questions';
+//import { getQuestions } from '../features/questionsSlice';
 
 // eslint-disable-next-line react/prop-types
 export const ViewQuestions = ({ selectedTopic }) => {
   const [selectedQuestion, setSelectedQuestion] = useState('');
   const dispatch = useDispatch();
-  const topicType = useSelector((state) => state.questionType);
+  const topicType = useSelector((state) => state.questionType.category);
 
-  useEffect(() => {
-    dispatch(getQuestions(selectedTopic));
-  }, [selectedTopic, topicType]);
+  // useEffect(() => {
+  // dispatch(getQuestions(selectedTopic));
+  // }, [selectedTopic, topicType]);
 
   const questions = useSelector((state) => state.questions);
   let arr = [];
