+++ client/src/features/store.js
@@ -1,12 +1,19 @@
 import { configureStore } from '@reduxjs/toolkit';
-import questionTypeReducer from './questionType';
-import topicReducer from './topic';
-import questionsReducer from './questions';
+import questionTypeReducer from './questionTypeSlice';
+import topicReducer from './topicSlice';
+import questionsReducer from './questionsSlice';
+import { questionsApi } from '../services/questionsApi';
+import { setupListeners } from '@reduxjs/toolkit/query';
 
 export const store = configureStore({
   reducer: {
     questionType: questionTypeReducer,
     topics: topicReducer,
-    questions: questionsReducer
-  }
+    questions: questionsReducer,
+    [questionsApi.reducerPath]: questionsApi.reducer
+  },
+  middleware: (getDefaultMiddleware) =>
+    getDefaultMiddleware().concat(questionsApi.middleware)
 });
+
+setupListeners(store.dispatch);
